feat: SolvedAC rateLimiter 추상화 및 테스트 코드 작성

+++ src/main/java/kr/solta/adapter/solvedac/SolvedAcRestClient.java
@@ -52,7 +52,7 @@ public class SolvedAcRestClient implements SolvedAcClient {
     }
 
     @Override
-    public List<SolvedAcProblemResponse> lookupProblems(final List<Integer> problemIds) {
+    public List<SolvedAcProblemResponse> lookupProblems(final List<Long> problemIds) {
         if (problemIds.isEmpty()) {
             return Collections.emptyList();
         }
