+++ src/main/java/kr/solta/application/required/dto/SolvedAcProblemResponse.java
@@ -3,7 +3,7 @@ package kr.solta.application.required.dto;
 import java.util.List;
 
 public record SolvedAcProblemResponse(
-        int problemId,
+        long problemId,
         String titleKo,
         int level,
         List<SolvedAcTagResponse> tags
